coredns: readiness- and livenessProbe tweaks
The readinessProbe was delaying the startup of coredns greatly. This
change brings startup times from ~15s to ~2-3s when testing locally. A
reduction of ~80%.

The livenessProbe would also take potentially more than a minute to
be declared unhealthy and restarted which is now tweaked down to about
30 seconds

For more information about the probes, see: https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/#configure-probes