Update building documentation for macOS
Problem: The k3s build scripts rely on GNU utils, like sed, so when
these scripts are run on a macOS environment they don't work as
expected.

Solution: Update the documentation to point out this scenario and
suggest the reader to spin up the Linux virtual machine defined on
the Vagrantfile within this project.

Signed-off-by: Jonnatan Jossemar Cordero <jonnatan.cordero@suse.com>