install.sh shebang portability using env
Use `#!/usr/bin/env sh` instead of `#!/bin/sh` to increase portability.

Use `#!/usr/bin/env sh` instead of `#!/bin/sh` to increase portability.

For more information see _[Why is it better to use "#!/usr/bin/env NAME" instead of "#!/path/to/NAME" as my shebang?](https://unix.stackexchange.com/questions/29608/why-is-it-better-to-use-usr-bin-env-name-instead-of-path-to-name-as-my)_

Signed-off-by: Carlo Corradini <carlo.corradini98@gmail.com>