Add ability to specify local storage class name
Problem: The default local-path storage class name is hard-coded to local-path. In certain scenario it is desired to use a default class with a different name.
It is possible to achive this today by disabling the default manifest and then defining a new manifest with the desired name. This leads to unnecessary overhead for ensuring that all storage class attributes remain in sync, especially in airgap scenario.

Solution:
The solution is to add a config option to 'k3s server' called default-local-path-storage-classname that is then used in the local-path manifest to render the k8s artifacts with the desired name.

Tests:
k3s server -h | grep default-local-path-storage-class-name