Bump remotedialer to fix deadlock on connection close
Fixes an issue where remotedialer clients may run into a deadlock when closing connections. This prevents the client from reconnecting to the server, and as the Close function has deadlocked, any health-checks that rely on checking remotedialer connection state will continue to pass as it claims to still be connected.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>