tunnel: handle pod IP reuse (#13212)
* tunnel: handle pod IP reuse

a valid tunnel/session may be deleted when an IP is reused while a
Complete pod (for example a job) was using that IP but is being gc'ed.

This causes timeouts to webhooks after directDial is attempted because
session was removed.

Solution is to track the owner of the IP and delete the entry only when
the the owner pod is deleted.

Signed-off-by: Julian Vassev <jvassev@gmail.com>
(cherry picked from commit 913005658c2cbb261ed14f541cdba99905dcf084)
Signed-off-by: Brad Davidson <brad.davidson@rancher.com>